Install latest driver and firmware for nexmon during build.

+++ stage3/03-nexmon/01-run-chroot.sh
@@ -2,8 +2,17 @@
 
 cd /home/pi
 wget https://kali.download/kali/pool/non-free-firmware/f/firmware-nexmon/firmware-nexmon_0.2_all.deb
-wget https://github.com/jayofelony/brcmfmac-nexmon-dkms/releases/download/v6.12.2%2Bndevfix3/brcmfmac-nexmon-dkms_6.12.2%2Bndevfix3_all.deb
+
+NEXMON_REPO="jayofelony/brcmfmac-nexmon-dkms"
+latest_tag=$(wget -qO- "https://api.github.com/repos/${NEXMON_REPO}/releases/latest" | sed -n 's/.*"tag_name": *"\(v[^"]*\)".*/\1/p')
+if [ -z "$latest_tag" ]; then
+  echo "could not determine latest nexmon driver release" >&2
+  exit 1
+fi
+latest_ver="${latest_tag#v}"
+nexmon_deb="brcmfmac-nexmon-dkms_${latest_ver}_all.deb"
+wget "https://github.com/${NEXMON_REPO}/releases/download/${latest_tag//+/%2B}/${nexmon_deb}"
 
 apt-get remove -y firmware-brcm80211
-apt-get install -y /home/pi/firmware-nexmon_0.2_all.deb /home/pi/brcmfmac-nexmon-dkms_6.12.2+ndevfix3_all.deb
-rm -r firmware-nexmon_0.2_all.deb brcmfmac-nexmon-dkms_6.12.2+ndevfix3_all.deb
+apt-get install -y /home/pi/firmware-nexmon_0.2_all.deb "/home/pi/${nexmon_deb}"
+rm -r firmware-nexmon_0.2_all.deb "${nexmon_deb}"
